+++ dart/lib/lipimala.dart
@@ -6,5 +6,6 @@ export 'deva_gujr_converter.dart';
 export 'latn_iast_to_deva.dart';
 export 'latn_iast_to_gujr.dart' hide RomanizationProfile;
 export 'latn_iast_transcription.dart' hide RomanizationProfile;
+export 'list_converters.dart';
 export 'transliteration_core.dart';
 export 'transliteration_result.dart';