+++ javascript/src/index.js
@@ -8,3 +8,4 @@ export * from './example-latn-iast.js';
 export * from './example-deva.js';
 export * from './example-gujr.js';
 export * from './deva-gujr-converter.js';
+export * from './list-converters.js';